Adjustments
The source from which we are taking bootstrap from is mainly approached
from a pure HTML approach.  Since Markdown spits out a fairly plain and
simple HTML, it can be difficult to apply this CSS how it was meant to,
this is an attempt to bring it closer to the original source.  Probably
still needs a bit of work.